Update NetConduit package to version 1.2.3
Updated the `NetConduit` package version from `1.2.2` to `1.2.3` in the following project files:
- `NetConduit.Ipc.csproj`
- `NetConduit.Quic.csproj`
- `NetConduit.Tcp.csproj`
- `NetConduit.Udp.csproj`
- `NetConduit.WebSocket.csproj`

The update applies to the `<PackageReference>` element under the condition where `$(UseLocalNetConduit)` is not set to `true`. This ensures that the updated version of the `NetConduit` package (1.2.3) is used when not relying on a local project reference. The change likely incorporates new features, bug fixes, or improvements introduced in version `1.2.3`.